avoid creating duplicate remote data objects
we were counting on scrapy on avoid duplicate requests and it does this just
fine, but we were creating a remote data object every time we encountered a
valid link. Since links to pages might pop up multiple times, the logic was
updated so we only create a rdo once.